SOW-GEN/requirements.txt
openai
streamlit
pypdf
xhtml2pdf
python-dotenv
